+++ modules/credentials/src/CredentialRole.ts
