+++ modules/credentials/src/services-example.ts
@@ -1,12 +1,19 @@
 import { AgentMessage } from '@aries-framework/core'
-import { ProposeCredentialOptions, CredentialRecord, ProtocolVersion } from './interfaces'
+import { ProposeCredentialOptions, AcceptProposalOptions } from './interfaces'
+import { CredentialProtocolVersion } from './CredentialProtocolVersion'
+import { CredentialExchangeRecord } from './CredentialExchangeRecord'
+
+interface CredentialServiceCreateReturnType { 
+  record: CredentialExchangeRecord;
+  message: AgentMessage
+}
 
 interface CredentialService {
-  createProposal(proposal: ProposeCredentialOptions): Promise<{ record: CredentialRecord; message: AgentMessage }>
+  createProposal(proposal: ProposeCredentialOptions): Promise<CredentialServiceCreateReturnType>
 }
 
 class CredentialServiceV1 implements CredentialService {
-  public async createProposal(proposal: ProposeCredentialOptions): Promise<{ record: CredentialRecord; message: AgentMessage }> {
+  public async createProposal(proposal: ProposeCredentialOptions): Promise<CredentialServiceCreateReturnType> {
     // should only handle the proposal.credentialFormats.indy
     return {
       /** return record and message */
@@ -15,7 +22,7 @@ class CredentialServiceV1 implements CredentialService {
 }
 
 class CredentialServiceV2 implements CredentialService {
-  public async createProposal(proposal: ProposeCredentialOptions): Promise<{ record: CredentialRecord; message: AgentMessage }> {
+  public async createProposal(proposal: ProposeCredentialOptions): Promise<CredentialServiceCreateReturnType> {
     // should handle all formats in proposal.credentialFormats by querying and calling
     // its corresponding handler classes.
 
@@ -26,18 +33,26 @@ class CredentialServiceV2 implements CredentialService {
 }
 
 export class CredentialModule {
-  private getService(protocolVersion: ProtocolVersion) {
+  private getService(protocolVersion: CredentialProtocolVersion) {
     const serviceMap = {
-      [ProtocolVersion.V1_0]: CredentialServiceV1,
-      [ProtocolVersion.V2_0]: CredentialServiceV2,
+      [CredentialProtocolVersion.V1_0]: CredentialServiceV1,
+      [CredentialProtocolVersion.V2_0]: CredentialServiceV2,
     }
     return new serviceMap[protocolVersion]()
   }
 
-  public async proposeCredential(options: ProposeCredentialOptions): Promise<CredentialRecord> {
+  public async proposeCredential(options: ProposeCredentialOptions): Promise<CredentialExchangeRecord> {
     const service = this.getService(options.protocolVersion)
     const { record, message } = await service.createProposal(options)
     this.messageSender.send(message)
+    return record
+  }
+
+  public async acceptProposal(options: AcceptProposalOptions): Promise<CredentialExchangeRecord> {
+    const credentialRecord = await this.credentialRepository.getById(options.credentialRecordId)
+    const service = this.getService(credentialRecord.protocolVersion)
+    const { record, message } = await service.acceptProposal(credentialRecord)
+    await this.messageSender.send(message)
     return record
   }
 }
